requirements.txt
aiogram==3.10.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
redis==5.0.1
pydantic==2.5.0
alembic==1.13.1
